safari-600.7-branch/LayoutTests/editing/spelling/spelling-exactly-selected-word-expected.txt

Spelling marker should remain when the user selects the misspelled word exactly. To test manually, type 'wellcome home.', select 'wellcome'. The test succeeds when 'wellcome' has spelling marker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.markerCountForNode(destination.childNodes[0], "spelling") became different from 0
PASS spellingMarkerRange.toString() is "wellcome"
PASS window.getSelection().toString() is ""
PASS window.getSelection().toString() is "wellcome"
PASS spellingMarkerRange.toString() became "wellcome"
PASS successfullyParsed is true

TEST COMPLETE
wellcome home.
